⚒️ Refactor: 세션id를 위한 파일구조변경
